refactor: changes frontend folder structure, created readme.md for frontend
