+++ frontend/README.md
@@ -1,21 +1,44 @@
-# React + TypeScript + Vite + shadcn/ui
+# Faculty Accomplishment Tracker
 
-This is a template for a new Vite project with React, TypeScript, and shadcn/ui.
+A structured reporting tool designed to streamline the recording, organization, and documentation of faculty achievements. This system acts as a bridge between the existing ISIP system and the PBMS, eliminating the need for department chairs to manually re-encode faculty data by providing customizable, PBMS-compatible data exports.
 
-## Adding components
+## Key Features
+* **Streamlined Submissions:** Faculty can easily input monthly accomplishments (teaching, research, extension) and upload supporting documents.
+* **Approval Workflow:** Centralized dashboard for department chairs to review, validate, and approve reports.
+* **PBMS Integration:** Export validated faculty data directly into PBMS-compatible CSV or Excel (XLS) formats.
 
-To add components to your app, run the following command:
+## Frontend Tech Stack
+* **Framework:** React
+* **State Management:** Zustand
+* **Data Fetching:** TanStack Query (React Query)
+* **Routing:** React Router
+* **Forms & Validation:** React Hook Form + Zod
+* **UI Components:** shadcn/ui
 
-```bash
-npx shadcn@latest add button
+## Getting Started
+
+### Prerequisites
+Make sure you have [Node.js](https://nodejs.org/) (v16 or higher) and `npm` installed on your machine.
+
+### Installation & Running the App
+
+1. **Clone the repository**
+   ```bash
+   git clone <repository-url>
+   cd cmsc127-project
+
+2. **Install dependencies**
+   ```bash
+   npm install
+
+3. **Start the development server**
+   ```bash
+   npm run dev
+
+4. **Open the app**
+Open your browser and navigate to the local server address (default is usually port 5173 for Vite/React):
+```text
+http://localhost:5173
 ```
 
-This will place the ui components in the `src/components` directory.
 
-## Using components
-
-To use the components in your app, import them as follows:
-
-```tsx
-import { Button } from "@/components/ui/button"
-```
