+++ frontend/src/components/dropzone-demo.tsx
@@ -1,4 +1,3 @@
-
 import { Dropzone, DropzoneContent, DropzoneEmptyState } from '@/components/ui/dropzone'
 import { useSupabaseUpload } from '@/hooks/use-supabase-upload'
