+++ frontend/src/features/forms/form-test/test-form.tsx
@@ -1,6 +1,6 @@
 import { formSchema, type FormValues } from "@/features/forms/form-test/test-schema"
 import { formFields } from "@/features/forms/form-test/test-config"
-import { DynamicForm as FormRenderer } from "@/components/forms/dynamic-form/dynamic-form"
+import { DynamicForm as FormRenderer } from "@/features/forms/dynamic-form/dynamic-form"
 
 
 export default function TestForm() {
